cognito-auth-package new/lambda-auth-authorizer/requirements.txt
python-jose[cryptography]==3.3.0
cryptography==42.0.2